// src/gen/endpoints/user/user.zod.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * STORE ME API
 * Hệ thống lưu trữ file cho phép người dùng upload, download và quản lý file của mình.

 * OpenAPI spec version: 1.0 beta
 */
import {
  z as zod
} from 'zod';


/**
 * @summary Lấy User theo id
 */
export const getUserByIdParams = zod.object({
  "id": zod.number()
})

/**
 * @summary Cập nhật User theo id
 */
export const updateUserByIdParams = zod.object({
  "id": zod.number()
})

export const updateUserByIdBodyUsernameRegExp = new RegExp('^[a-zA-Z0-9]{3,20}$');


export const updateUserByIdBody = zod.object({
  "username": zod.string().min(1).regex(updateUserByIdBodyUsernameRegExp),
  "email": zod.string().min(1),
  "roleId": zod.number()
})

/**
 * @summary Xóa User theo id
 */
export const deleteUserByIdParams = zod.object({
  "id": zod.number()
})

/**
 * @summary Cập nhật trạng thái người dùng
 */
export const updateUserStatusParams = zod.object({
  "id": zod.number()
})

export const updateUserStatusQueryParams = zod.object({
  "status": zod.enum(['UNVERIFIED', 'ACTIVE', 'BLOCKED', 'DELETED'])
})

/**
 * pagination + filter với turkraft/springfilter
 * @summary Lấy danh sách User
 */
export const getAllUserQueryPageDefault = 0;
export const getAllUserQueryPageMin = 0;
export const getAllUserQuerySizeDefault = 20;

export const getAllUserQueryParams = zod.object({
  "page": zod.number().min(getAllUserQueryPageMin).optional().describe('Zero-based page index (0..N)'),
  "size": zod.number().min(1).default(getAllUserQuerySizeDefault).describe('The size of the page to be returned'),
  "sort": zod.array(zod.string()).optional().describe('Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'),
  "filter": zod.string().describe('Filter by field with turkraft/springfilter. Example: filter=field1==\'value1\' and field2>5')
})

/**
 * @summary Tạo mới User
 */
export const createUserBodyUsernameRegExp = new RegExp('^[a-zA-Z0-9]{3,20}$');


export const createUserBody = zod.object({
  "username": zod.string().min(1).regex(createUserBodyUsernameRegExp),
  "email": zod.string().min(1),
  "roleId": zod.number()
})

/**
 * @summary Xóa nhiều User theo danh sách id
 */
export const deleteUserBulkBodyItem = zod.number()
export const deleteUserBulkBody = zod.array(deleteUserBulkBodyItem)